major contactform update, deleted partial and contact page, moved to index for server and client validation
