--- MillePortfolio/Pages/Contact.cshtml
@@ -1,4 +1,0 @@
-﻿@page
-@model MillePortfolio.Pages.ContactModel
-@{
-}
