--- MillePortfolio/Pages/Shared/_ContactForm.cshtml
@@ -1,30 +1,0 @@
-﻿<div class="contact-form-start container">
-    <h2 class="shadow">Kontakta mig</h2>
-    <form method="post" asp-page="/Contact">
-        <div class="form-row">
-            <div class="d-flex justify-content-between ">
-                <div class="form-group col-md-6 top-left-contact">
-                    <label for="inputName"></label>
-                    <input type="text" class="form-control shadow" id="inputName" name="Contact.Name" placeholder="Ditt namn">
-                </div>
-                <div class="form-group col-md-6 top-right-contact">
-                    <label for="inputEmail"></label>
-                    <input type="email" class="form-control shadow" id="inputEmail" name="Contact.Email" placeholder="Din email">
-                </div>
-            </div>
-        </div>
-        <div class="form-group">
-            <label for="inputSubject"></label>
-            <input type="text" class="form-control shadow" id="inputSubject" name="Contact.Subject" placeholder="Ämne">
-        </div>
-        <div class="form-group">
-            <label for="inputMessage"></label>
-            <textarea class="form-control shadow" id="inputMessage" name="Contact.Message" rows="3" placeholder="Ditt meddelande"></textarea>
-        </div>
-        <button type="submit" class="btn btn-secondary equal-width mt-3 mx-auto d-block shadow contact-btn">Skicka <i class="fa-solid fa-arrow-right"></i></button>
-    </form>
-</div>
-
-@section Scripts {
-    <partial name="Shared/_ValidationScriptsPartial" />
-}
